requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
pip==23.0.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.10.0
urllib3==1.26.15
wsproto==1.2.0
